tty: msm_serial_hs: Set kernel Fmax if freq is too high at probe
On some devices the bootloader may set a higher-than-normal
frequency on the uart port.
This probably happens when the bootloader tries to close it
prior jumping to the Linux kernel due to some bug.
In these cases we definitely want to force the clock to have
the maximum allowed frequency declared in the clock table
but, of course, not higher.