LEGACY-PORT: msm: kgsl: Use del_timer() where appropriate
The real use of del_timer_sync() comes in cases where
we want to free shared resources, shared with the timer
handler. In those cases, we want to make sure that if
the timer handler was running on another cpu, it must
finish executing before we go ahead with free'ing the
shared resources. So basically, to avoid a use-after-free.

However, in this code path, we are not going to destroy
the shared resources. So, here we don't need to wait
for timer to finish, we can return immediately.

Change-Id: I9bf7a6b215f5ba86049f5df6bc82eed6958cf89d
Signed-off-by: Harshdeep Dhatt <hdhatt@codeaurora.org>