gpu: drm: sde: Fix interrupts list for pre-AD4.x targets
Targets that have an older revision of the AD block haven't got
the additional interrupts and trying to set them will produce an
HW fault: create a list of non-AD4 targets and exclude these
interrupts for them.

NOTE: Everyone hates ifdefs, but this is the only way to prevent
a porting nightmare in the future.