rqbalance: Schedule CFL setting on hotplug if cluster unavailable
If a cluster is totally down, it is impossible to set a cpufreq
policy on it.

Change the logic so that, if the cluster is currently up we set
the limits right in the sysfs handler but, if the cluster is
unavailable, schedule the setting for the next upcore event and
artificially produce a non-blocking upcore event on the
navailable cluster: this will make us get into the newly
introduced hotplugging notifier that will take care of updating
the policy (thus setting the frequency limits) as soon as the
cluster will come up.