LEGACY-PORT: msm: kgsl: Sync non io-coherent dma buffers on attach to GPU
Sync the dma buffer when it is attached to the GPU device.
This is necessary to make sure any data in the buffer
is flushed to memory at attach time. Skipping it can
lead to coherency problems between CPU and GPU. The sync
can be skipped for IO-coherent DMA buffers.

Change-Id: I06fa54822688b1834e83a4e137c95d1c681300fb
Signed-off-by: Sushmita Susheelendra <ssusheel@codeaurora.org>
Signed-off-by: Angelo G. Del Regno <kholk11@gmail.com>