input: ts: dsx_td4322: Add PM notifiers for DRM
In case that we are using a DRM driver instead of a FBDEV one,
we wouldn't be able to suspend the driver properly, due to hw
hacks in the target devices.

Add a DRM notifier to catch blank and unblank events from it
and promptly suspend and resume the driver when needed.