pinctrl: msm: Remove all IRQCHIP flags from msmgpio
First of all, we want to remove the IRQCHIP_SKIP_SET_WAKE flag,
because we definitely do want to set IRQ wake, especially before
the system enters mem sleep (suspend - deep sleep), so that we
get certain GPIOs set up to fire an interrupt that is able to wake
up the system.

Also, remove the IRQCHIP_MASK_ON_SUSPEND flag, so that we are sure
that we are not discarding any operation (any setting) done before
perhaps using the irq_set_wake handler.

This ensures that GPIO controller can fire interrupts that are
able to wakeup the system during suspend.