rqbalance: Provide compatibility with CPU isolation
This driver was meant to be used with CPU hotplugging and
does not support CPU isolation.

Provide compatibility with both isolation and hotplugging by
switching the mask of "available CPUs" to either a cpumask
of online CPUs or non-isolated CPUs at governor start time.

Of course, the algorithm was executing for_each_online_cpu,
but it was really meaning "for each currently usable core"
and the definition of "usable core" is either one that is
either plugged in or not isolated.