rqbalance: Use the available CPUs mask for CPU_UPCORE compat
The CPU_UPCORE case will be used to unisolate a CPU in case
the current performance control way is sched isolation
instead of hotplugging.
This means that we fall into in both hotplug and isolation
cases and here we must select the next unavailable CPU, not
limited to it being offline.