clk: qcom: clk-hfpll: Update MN registers when setting PLL rate
The PLL's M/N registers were not updated when setting its rate:
this is important because, in the rare event of ESD, these
registers can get corrupted and we may get an over/under clock
of the PLL output (and whatever sources it).

While at it, update the L register setting to use the convenient
regmap_update_bits function instead of regmap_write.