mmc: core: Add cmd6 time quirk for HYNIX eMMCs and other SoMC hacks
Even if it looks like working correctly with the default
parameters read from ext_csd, this is required for best
performance, at least on SoMC Tone platform (MSM8996).

Also, add a hack for the sa_timeout because of bad CSD on
SoMC Nile and Ganges platforms and enable the management
of the MMC_CAP2_FULL_PWR_CYCLE for SoMC Loire and Tone.

As last, disable enhanced strobe for SoMC Loire only, as
the platform's card supports it but the SDHCI controller
on these boards to not support it fully, leading to various
errors and reinitializations of the eMMC card during the
kernel life.