msm: kgsl: Add property KGSL_PROP_GPU_FORCE_ON and k4.9 binaries compat
Add property KGSL_PROP_GPU_FORCE_ON to allow user to force
GPU state to ON. Also, to do this, a modification is strictly
required on the current kernel (4.14) driver:
add a configuration option (CONFIG_QCOM_KGSL_OLD_BINARIES_COMPAT)
to retain compatibility with the kernel 4.9 associated userspace
binary companion.

To clarify that, here is reported the same text as the help that
I wrote for the configuration option:

The old userspace binaries may get/set some KGSL property
which definition is different from the new binaries.
For example, KGSL_PROP_SPEED_BIN (from 4.14) is 0x25, but
KGSL_PROP_GPU_FORCE_ON (unsupported in 4.14) corresponds
to the same number on k4.9 adreno userspace binaries.
Say Y here if you want to run old adreno binaries, for a
reason or another.