clk: qcom: alpha-pll: Add support for PLL non-dynamic update
This commit adds proper support for PLLs that do not support
dynamic programming (some legacy Alpha PLLs haven't got the
dynamic_update bit set) and refreshes support for dynamic
update PLLs.

In the case of non-dynamic update, ensure that we disable
and re-enable the PLL everytime we want to change its rate
because this is the only way to do it.