irqchip: arm-gic-v3: Implement GIC save and restore API for SDM845
On the SDM845 SoC we require to save and restore SPI to
keep functionality when resuming the GIC as it would lose
the configuration when sent to sleep.

This is a port of kernel 4.9 commit
b179a8e1e9f8c085f64fb0b60f1244a882fad01e

Provide api to save/restore GICD state for SPIs. SPI
configuration is restored for GICD_ICFGR, GICD_ISENABLER,
GICD_IPRIORITYR, GICD_IROUTER registers. Following is the
sequence for restore:

1. For SPIs, check whether any of GICD_ICFGR, GICD_ISENABLER,
   GICD_IPRIORITYR, GICD_IROUTER, current configuration is
   different from saved configuration.

For all irqs, with mismatched configurations,

2. Set GICD_ICENABLER and wait for its completion.

3. Restore any changed GICD_ICFGR, GICD_IPRIORITYR, GICD_IROUTER
   configurations.

4. Set GICD_ICACTIVER.

5. Set pending for the interrupt.

6. Enable interrupt and wait for its completion.