global: Wire up Qualcomm's ugly audio techpack
This wires up Qualcomm's ugly audio techpack folder to the
rest of the kernel.

Why they keep the audio stuff in a completely separated GIT
repository is not really excusable, since they are kernel
drivers and they belong to the kernel... but since that's
how it gets distributed, that's the only way that I have
to integrate it into the kernel.

P.S.: Building audio as modules only is not an excuse!!!
      This is an embedded platform!!!