arm64: mm: Fix building for !STRICT_KERNEL_RWX && !STRICT_MODULE_RWX
When these two kernel configuration options are not defined
the build would fail due to AArch64 code requiring the
rodata_enabled variable in two internal functions.

To fix this issue, define a static rodata_enabled privately
inside of mmu.c when these configuration options are not
enabled and initialize it to false.