scsi: ufs: Fixup Sony quirks for kernel 4.14
There were some API changes that were missed, plus change this
and that to make the code more readable.

Also, change the definition from a per-platform ifdef to
a header checking if the platform is in the ifdef list and
defining a single UFS_TARGET_SONY_PLATFORM to activate the
codepaths.
This decreases filesize and enhances the ease of porting.