clk: qcom: gcc-sdm660: Add a fake bi-tcxo clock for compatibility
On kernel 4.14 the 14nm MDSS PLL is being used for SM6150 and some
other SoC, which are RPMh based: this made Qualcomm to change the
parent name to the RPMh's bi_tcxo, instead of doing it properly
and vary the clock parent name based on a property or based on the
SoC or based on a DT binding for the right XO clocksource.

Since there is no time to do it properly on my side right now,
this will be the solution that we'll temporarily use, which is for
sure dirty, but effective: just add a fake bi_tcxo clock, which
parent is the xo, to satisfy the dependency of the MDSS PLL clocks.