tty: serial: msm_serial: Do not rely on zero setting for early TX restore
The early TX pin restore code was relying on a less than 0x1000000
setting to get disabled: this is generally bad and totally wrong.

For this reason, add another configuration option so that we
totally disable the early UART TX restore code, which is anyway
used only for early kernel porting, since the normal way to
restore the pin sort-of-early is with the restore_msm_uart
boot parameter.