usb: dwc3-msm: Fix registering USB-DropT for overcurrent protection
The overcurrent protection happens only on the PDPHY and we cannot
register it on other extcons because that's not available.

As a sort of solution, try to register it at every iteration of
the extcon notifier registration. We require to succeed once,
as multiple OCP notifications on the same PDPHY on the same
DWC3 controller are not useful: we just need one to protect the
hardware.

P.S.: It's dirty, but also the only way to do it since the
edev name is not exposed to the drivers (nor struct extcon_dev).