arm64: DT: smmu-660: Do not read the ACTLR register on Q6 SMMU
The new security scheme, appeared after a bootloader update, locks
out the ACTLR register on the Q6 SMMU, producing a lockup when the
HLOS tries to read it because the hypervisor generates a security
halt.

Stop reading it so that we can support booting Linux on the newest
bootloader revision.