soc: qcom: system_stats: Fix RPM RAM memory regions names and meanings
The RPM RAM regions were wrongly named in there.
The first one refers to the RPM Data RAM iospace and the second one
refers to the Message RAM. The region names were basically swapped.

Also, fix the retrieval of the Data RAM, so that we can declare just
the relevant address in DT, instead of declaring the RPM base and
the RPM Data RAM base, while only using the latter.

An example node declaration would be:
	rpm_data_ram: memory@0x200000 {
		compatible = "qcom,rpm-data-ram";
		reg = <0x290000 0x1000>;
	};

	qcom,system-stats {
		compatible = "qcom,system-stats";
		qcom,rpm-data-ram = <&rpm_data_ram>;
		qcom,rpm-msg-ram = <&rpm_msg_ram>;
		qcom,masters = "APSS", "MPSS", "ADSP", "CDSP", "TZ";
	};