cpufreq: interactive: WORKAROUND: Delete timer to address race condition
There's a race condition happening in the CPU hotplugging path:
sometimes some timer doesn't get canceled because some core goes
down before this happens.

As a workaround to stop crashing the kernel every now and then,
make sure that the timer is not pending before rescheduling it.

This is a bad solution and has to be solved in the cpuhp API
instead, but it's something to keep things up for now.