kconfig: lxdialog: Stub out pkg-config for Android Q build system
The Android build system explicitly disallows us from calling
pkg-config on the host system.
To address this poor issue, the only solution is to stub the
pkg-config call so that we find headers and libs to link in
another (the good old) way.

P.S.: I really don't like this, but there's no other solution.