clk: qcom: alpha-pll: Fix Trion PLL calibration on prepare ops
The Trion PLL shall do a power-on self-calibration to work,
which is performed on an output frequency of around 1300MHz.

The prepare function needs to start this calibration, but there
a typo was made: we were doing it by multiplying the current L
value of the PLL by XO instead of multiplying the CAL value.
This was pushing the PLL at the wrong frequency, in which it's
not possible to perform self-calibration.

Fix it by multiplying the right value.