gpu: msm: Fast forward Adreno driver to LA.UM.8.1.r1-11600-sm8150.0
This is a fast forward of the Adreno driver to the new 8.1.r1 branch
for Android 10 and it is necessary for good compatibility with the
newest userspace proprietary drivers for the Adreno GPUs.

Of course, this already contains the patches that were applied to
the previous branch Adreno driver in order to retain compatibility
with legacy SoCs.