cpuhp/rqbalance: Add kernel parameter to start in performance mode
rqb_start_perf starts the governor in "performance mode",
which means that the CPUs will have a big tendency to be
onlined at boot: this speeds up the boot process, as many
userspaces do spawn few very power hungry threads at boot.

This performance mode will be discarded as soon as the
userspace will take control of the parameters for this
driver.