arm64: DT: SDM630/660: Guard msm shared memory region
If guard_memory is set, then the shared memory region will be
guarded by SZ_4K at the start and at the end.

This is needed to overcome the XPU limitation on few MSM HW,
so as to make this memory not contiguous with other allocations
that may possibly happen from other clients in the system.