iommu: io-pgtable-arm: Always program TTBR1 to avoid S2 faults
On TTBR1 we may see address translation occurring even if it is
disabled for some reason (and that's probably an errata of some
kind for the ARM IOMMUs).
This leads to a S2 fault, which may trigger a hypervisor fault
on some Qualcomm SoCs featuring some security configurations.

To stop the hypervisor from generating a fault (thus rebooting
the HLOS), always allocate a page on the TTBR1 which, in case
the TTBR1 pgtable quirk is not declared on the currently used
page table, will be a zero page - but it will be valid.

In this case, if we have a valid zero page on TTBR1, when the
address translation occurs, the IOMMU will still throw a fault,
but this time it will be S1, which won't trigger the hypervisor.