gpu: drm: panel_color_manager: Do not fail merge on missing color table.
Breaking out of atomic_set_property without forwarding any call results
in a black screen on Bahamut. somc_panel_update_merged_pcc_cache returns
-EINVAL because this panel does not have a color table.

Ignore this failure and instead rely solely on the state of the cache.
If the merge cache is valid: use that. If not, use the transformation
requested through atomic_set_property directly.

Aside from that, clean up somc_panel_update_merged_pcc_cache where this
issue resides in the first place: missing color tables are the perfect
opportunity to fall back to the requested system calibration.
Also invalidate the cache upon entering the function. If breaking out
for any failure reason the cache might still be valid yet outdated;
prevent this from being used and rely on the caller to fall back on the
sole unmerged PCC to be used.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>