drm: msm: somc_panel: Send AOD OFF only if coming from AOD
It is wrong (and useless, actually) to send the AOD OFF command
while we are powering up the display from display-off state.

Besides that, some DriverIC (Samsung, for instance) really does
dislike getting this command when powering up the display from
a complete DPMS_OFF state, showing unesthetic corruption for
the first frame.