arm64: DT: SDM630/660: Properly configure for 4.14 LMH DCVS driver
The LMH DCVS driver in kernel 4.14 wants to be configured a little
differently from the one in 4.9, since the registers are mapped
in a different way inside the driver.
Also, now the modern driver features support for the old limits
management and we can use it just fine and without the (not critical)
errors that we used to just ignore before.