arm64: DT: SDM630/660: Use sdm660-class quirks for QCOM GICv3
On these SoCs it's needed to skip setting the wake bit on the
interrupts (as it's necessary to do that only on the MPM) and
it's necessary to mask the lines on suspend.

Setting the wake set of interrupts is not necessary here and
will lead to some issues like the unability of resuming the
AP due to modem interrupts (for example, receiving calls will
not wake up the device).