media: msm: rotator: Properly implement shared IOMMU context quirk
Some MDP versions (3.2.0, 3.3.0, found in SDM630/660) are configured
to share one IOMMU context between MDP and ROT devices: in these
cases we cannot totally free and remap the memory there from the
rotator driver because, in the event of a free, we would also be
freeing the MDP mapped space, creating havoc (as that's being used
for the display in general).

For this reason, this kind of configuration requires us to map
only once (in the mdp driver, actually) and just make use (and
invalidation) of the portion of memory associated to the rotator
hardware, hence we need to differentiate the handling for this.

This commit also assumes that only rotator hardware requiring
software timestamp needs the quirk (as, anyway, the mapping is
actually done only in that case).

So, define a new SDE_CAPS_SPLIT_CTX cap entry and assign it to
MDP versions 3.2.0, 3.3.0, then use this codepath only if the
cap is set, else, just go on with the regular code that assumes
that we have one iommu context for mdp and one for rotator.