drm: msm: somc_panel: Implement HBM with safe handling
Implement HBM with safe handling for Samsung Electronics panels,
currently found in SoMC Kumano platform.

The High Brightness Mode (HBM) allows the panel to reach up to
20% more nits, but also provides some stress to the OLED matrix:
in order to avoid burn-in, the user will be allowed to stay in
HBM mode for a maximum of 10 minutes, with a cooldown time of
3 minutes.

If the user doesn't exit HBM (through auto-brightness in the
userspace, or through manual bightness setting), then this
driver will enable it for 10 minutes, disable it for 3, then
re-enable it for 10 - for up to 3 times.

Entering the HBM mode requires the userspace requesting MAXIMUM
brightness for the target panel.

This feature will not be enabled by default for all panels that
support the High Brightness Mode commands: these have to be
validated manually before deciding on its safety, so each panel
needs to declare in its device-tree configuration the property
"somc,set-hbm-usable", otherwise the automatic switch to HBM
will be made impossible.

P.S.: Remember, with great power comes great responsibility...