drm: msm: Fix split display configurations on MDP5 SoCs
Some MDP5 SoCs, like MSM8994, 8996, 8998, have the left LM at
LM0 and the right LM at LM2.

This driver is always expecting LM0 and LM1 for split display,
which is right when it comes to DPU1, but it's not guaranteed
to be like that on all MDP5.

To fix this behavior, add new DT configurations for the mixer
display preference:
- primary-split   - Forces adding split display but on a LM that
                    supports hosting a primary display
- primary-nosplit - Removes split display support bit from current LM
- split           - Forces adding split display bit on current LM
- nosplit         - Forces removing split display bit on current LM