cpufreq: interactive: Disallow starting as default governor
This governor requires slack timers, which are not initialized
at early boot, causing an early kernel panic.

For this reason, remove the option and relative code to set it
as default governor, so that we're *forced* to set it up once
the userspace has booted.
Anyway, this governor requires device-specific tuning, or it
won't reall work.

TODO: Find a replacement for this thing......