clk: qcom: alpha-pll: Fix PLL enabled check for Alpha PLL
The Alpha PLL set rate function needs to check for clk enabled
to decide whether to get the PLL up after setting its rate, or
to wait for a dynamic update, or to "just set it and fly away":
there was a bad bad bad typo in here, where we were checking for
a non-NULL function to check if the PLL is enabled instead of
checking for the clock being enabled.

Obviously, the initial intention was the latter.