clk: qcom: gcc-msm8998: Write static configuration for GPLL0-4
The GPLLs 0 to 4 on MSM8998 are used as fixed rate PLLs but, in
any case, it may happen for one of the GPLLs to lose the L
calibration value (CAL_L): in this case, we have no configuration
to restore, so the SoC will crash because of either a hypervisor
fault or a unrecoverable hardware fault.

Add the configuration for GPLL0-4, so that we stop relying on
bootloader being kind with us and so that we can recover from
a PLL configuration loss.