clk: qcom: alpha-pll: Constify alpha_pll_config in clk_alpha_pll
In struct clk_alpha_pll it is safe to constify the alpha_pll_config
because this configuration never changes during the boot life time.

This configuration structure is used to recover the PLL to safe
values in the event that the PLL loses its calibration or even
all of its values.