clk: qcom: gcc-msm8998: Do not register GPLLs 1,2,3
These GPLLs are unused on this SoC and there is no specification
to follow to include usage of any of these in any clock.
Besides that, they are left off from the bootloader, so it's
useless to even have them here.

Ah, by the way... on Sony Yoshino platform, if we ever try to
touch one of these three GPLLs, we get an immediate hypervisor
fault and a not-so-nice device reboot.