platform: msm: ipa_v3: Set right parameters for rmnet on legacy
Legacy SoCs, such as MSM8998 and SDM630/636/660/845 need different
parameters for the rmnet in order to reach the maximum throughput
and best connection latency: set them through a ifdef, even if it
is also possible to set most of them (but one) through cmdline
parameters. After all, we're varying a "default" value, anyway.