msm: gsi: Free memory and zero refcount in MSM8998 workaround path
If we are running the MSM8998 workaround, the memory may be still
allocated even though the firmware has restarted and lost the GSI
channels states, since that's kernel-allocated memory.

Also, if the channel is not allocated in firmware, then we *have*
to set the channel refcount to zero, otherwise IPA will fail to
deinitialize, starting a glorious kernel panic scenario.