Initial port of Cheeseburger device tree from Kernel 4.4
The only node disabled are the 3 LEDs (red, green and blue).
At the moment, when the kernel is loaded nothing happens and after ~3
seconds the device reboots itself.

The cheeseburger_android_defconfig defconfig is added starting from the
aosp_yoshino_maple_defconfig defconfig. There is one difference: DRM is
disabled and 'standard' FB is enabled.

Change-Id: Ifa9e642b1c2a62739f38c1feeb41aa755396d0e9
Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>

# Conflicts:
#	include/dt-bindings/msm/power-on.h