scripts: Makefile.lib: dtc: add both "linux,phandle" and "phandle" properties
The dtc tools adds a "phandle" property to every node.
In the past, the "linux,phandle" was used instead and now it is
deprecated.
Kernel 4.4 still add it by default and the bootloader from OnePlus
expects it to be present.

With the '-H both' property, we can make sure to have both the
"linux,phandle" and the "phandle" properties to be compatible with the
bootloader from OnePlus.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>