techpack/audio: Delete Makefile that is already checked in to subrepo
This file is also checked in to the techpack-audio [1] repository and
causes unnecessary conflicts when it's managed by two git repositories
at the same time.  It's not necessary to replace it with a placeholder
to keep the folder structure alive as `repo` will create a directory
when syncing.

Unlike wlan-qc, no submodules are used for any of the techpack
repositories.

[1]: https://github.com/sonyxperiadev/kernel-techpack-audio

Signed-off-by: Marijn Suijten <marijn.suijten@somainline.org>