NFC: ese: add driver for the NXP P61 Secure Element chip
This is found in both OnePlus 5 and OnePlus 5T.
This is porting from kernel 4.4, with the only difference that we
include the "../../misc/pn553/pn553.h" file from a different location
(changed between kernel 4.4 and kernel 4.14).
May need some adjustment from the Android user space, but the kernel is
probing the device correctly on my OnePlus 5.

Also note that this is now dependent from the NFC_PN553_DEVICES driver,
as it uses some ioctls functions from there.

Change-Id: I5cb6313570f19283a839b90ef51670863346abcb
Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>