drivers: import changes to the QPNP SMB Battery Charger from kernel 4.4 to support Dash charge
This also removes support for the QPNP_SMBFG_NEWGEN_EXTENSION driver on
the fuel gauge gen3 and SMB2 battery charger. It is used only in devices
from Sony, that don't support Dash charge.

Ported from the kernel 4.4 for the OnePlus 5/5T (LineageOS).
Similarly as the porting of the Dash charge driver:
- change: (method) wake_lock -> __pm_stay_awake
- change: wake_unlock -> __pm_relax
- change: (struct) wake_lock -> wakeup_source
- change: wake_lock_init -> wakeup_source_init
- removed (commented): push_component_info

And more to build on 4.14.

Change-Id: I0106bed5f95ee5774ff9d0f0b36b3f7abe172eea
Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>