input: ts: sec_ts: Remove spam from event buffer empty
On some devices like Xperia 10 II this can happen dozens of times
just by touching the screen, remove the message and explain.

The event buffer is empty message is triggered by a bug since the event buffer isn't supposed
to be empty. However since the bug has no real impact so far and the message
is filling up the kernel logs we disable the message.

Fixes: sonyxperiadev/bug_tracker#738
Signed-off-by: Björn Bidar <bjorn.bidar@jolla.com>