regulator: rpm-smd-regulator: Reenable PIN_CTRL_MODE support for LDOA regulators
This was disabled with the introduction of set_load for new LDOA
regulators. In msm8998, however, we need this support for BT/ICNSS
LDOA regulators as they follow hardware pins for forcing HPM;
using only the general hpm_min_load default and the new set_load
function causes bluetooth to become unstable when turning off wifi
or turning on airplane mode.

Signed-off-by: ederekun <sedrickvince@gmail.com>